Move wallet to the SDK
Wallet is used quite heavily within the SDK for inputs, thus making it
apart of the SDK
