+++ apps/src/lib/cli/wallet.rs
@@ -7,8 +7,8 @@ use borsh::BorshSerialize;
 use color_eyre::eyre::Result;
 use itertools::sorted;
 use masp_primitives::zip32::ExtendedFullViewingKey;
-use namada::ledger::wallet::{DecryptionError, FindKeyError};
 use namada::sdk::masp::find_valid_diversifier;
+use namada::sdk::wallet::{DecryptionError, FindKeyError};
 use namada::types::key::*;
 use namada::types::masp::{MaspValue, PaymentAddress};
 use rand_core::OsRng;
