+++ shared/src/sdk/signing.rs
@@ -27,8 +27,6 @@ use zeroize::Zeroizing;
 use crate::ibc::applications::transfer::msgs::transfer::MsgTransfer;
 use crate::ibc_proto::google::protobuf::Any;
 use crate::ledger::parameters::storage as parameter_storage;
-pub use crate::ledger::wallet::store::AddressVpType;
-use crate::ledger::wallet::{Wallet, WalletUtils};
 use crate::proto::{MaspBuilder, Section, Tx};
 use crate::sdk::masp::{
     make_asset_type, ShieldedContext, ShieldedTransfer, ShieldedUtils,
@@ -42,6 +40,8 @@ use crate::sdk::tx::{
     TX_UNBOND_WASM, TX_UPDATE_ACCOUNT_WASM, TX_VOTE_PROPOSAL, TX_WITHDRAW_WASM,
     VP_USER_WASM,
 };
+pub use crate::sdk::wallet::store::AddressVpType;
+use crate::sdk::wallet::{Wallet, WalletUtils};
 use crate::sdk::{args, rpc};
 use crate::types::error::{EncodingError, Error, TxError};
 use crate::types::key::*;
